perf(tokenizer): add borrowed vocabulary fast path
