+++ docs/public-contract-1.0-candidates.md
@@ -13,6 +13,7 @@ The following surfaces are candidates for stabilization before the first stable 
 - `tokenize_fasta_records_reader`
 - `load_vocab_json`
 - `protein_20_vocab_tokens`
+- `protein_20_vocabulary`
 - `ProteinTokenizer` and `Tokenizer`
 - `ModelInput`, `ModelInputPolicy`, `PaddingPolicy`
 - `build_model_inputs_checked`
