+++ packages/rust/biors-core/src/lib.rs
@@ -33,10 +33,11 @@ pub use sequence::{
 };
 pub use tokenizer::{
     load_protein_20_vocab, load_vocab_json, protein_20_unknown_token_policy,
-    protein_20_vocab_tokens, summarize_fasta_records_reader, summarize_tokenized_proteins,
-    tokenize_fasta_records, tokenize_fasta_records_reader, tokenize_protein, ProteinBatchSummary,
-    ProteinTokenizer, SummarizedFastaInput, TokenizedFastaInput, TokenizedProtein, Tokenizer,
-    UnknownTokenPolicy, VocabToken, Vocabulary, PROTEIN_20_UNKNOWN_TOKEN_ID,
+    protein_20_vocab_tokens, protein_20_vocabulary, summarize_fasta_records_reader,
+    summarize_tokenized_proteins, tokenize_fasta_records, tokenize_fasta_records_reader,
+    tokenize_protein, ProteinBatchSummary, ProteinTokenizer, SummarizedFastaInput,
+    TokenizedFastaInput, TokenizedProtein, Tokenizer, UnknownTokenPolicy, VocabToken, Vocabulary,
+    PROTEIN_20_UNKNOWN_TOKEN_ID,
 };
 pub use verification::{
     stable_input_hash, verify_package_outputs, verify_package_outputs_with_observation_base,
